Crate regalloc3

A flexible and fast register allocator implementation designed to be embedded in an existing compiler code base.

This crate is compatible with #![no_std] and only requires alloc.

Usage

To use this register allocator in your project, you will need to define types which implement 2 traits.

You will need an implementation of RegInfo which describes the set of CPU registers available to the allocator and their properties. See the reginfo module documentation for more details. Generally, you will have a register description for each target architecture supported by your compiler.

You will also need an implementation of Function which describes the code on which the register allocator needs to run. See the function module documentation for more details.

The register allocator is invoked by creating an instance of RegisterAllocator and then calling RegisterAllocator::allocate_registers.

Once register allocation succeeds, it will return an Output which describes the registers that have been assigned to each instruction operand and the sequence of move/spill/reload instructions that need to be inserted between the original instructions. See the output module documentation for more details.

Reusing allocations

For performance reasons, the RegisterAllocator type don’t free temporary allocations after a call to RegisterAllocator::allocate_registers.

This allows register allocation to be run on multiple functions without the need for further calls to the memory allocator, which can be slow. If the memory usage is a concern or if you are finished compiling functions then you can simply drop RegisterAllocator to free all temporary memory.

Validation

When developing a new client of the register allocator, it is highly recommended to use the validation functions in debug_utils to ensure that the inputs to the register allocator are correct.

If inputs fail validation then the register allocator may panic, get stuck in an infinite loop or just produce garbage results.

However validation is relatively expensive so you may not want to have it enabled in production.

Note that even code that passes validation may cause the register allocator to return an error (RegAllocError). This is usually an indication of impossible constraints on an instruction.

Modules

debug_utils

Utility functions and types for debugging register allocation.

entity

The public API of the register allocator make heavy use of “entities”: these are newtype wrappers around integers which represent an index into an array.

function

The Function trait which describes the input function.

output

Output of register allocation.

reginfo

The RegInfo trait which describes the registers and register classes in the target ISA.

Structs

Options

Configuration options for the register allocator.

RegisterAllocator

Structure holding persistent memory allocations that can be reused across multiple invocations of the register allocator.

Stats

Statistics collected by the register allocator.

Enums

MoveOptimizationLevel

Controls how much optimization to perform after register allocation.

RegAllocError

Error returned by the register allocator if allocation is impossible.