regalloc3
Module debug_utils

Utility functions and types for debugging register allocation.

These are not needed for normal compilation, but are useful during development of both the register allocator itself and users of the register allocator.

Structs

ArbitraryFunctionConfig

Configuration options for GenericFunction::arbitrary_with_config.

ArbitraryRegInfoConfig

Configuration options for GenericRegInfo::arbitrary_with_config.

CostModel

Cost model to evaluate and compare the quality of different register allocation algorithms.

DisplayFunction

Wrapper around a type implementing Function that provides a Display implementation which dumps the function in a format that is both human-readable and machine-parseable.

DisplayRegInfo

Wrapper around a type implementing RegInfo that provides a Display implementation which dumps the register description in a format that is both human-readable and machine-parseable.

GenericFunction

A generic implementation of Function which can be constructed from an existing Function or parsed from a text representation.

GenericRegInfo

A generic implementation of RegInfo which can be constructed from an existing RegInfo or parsed from a text representation.

Functions

check_output

Verifies the output of the register allocator.

validate_function

Checks func to ensure it satisfies all of the pre-conditions required by the register allocator.

validate_reginfo

Checks reginfo to ensure it satisfies all of the pre-conditions required by the register allocator.