regalloc3::debug_utils
Function check_output

pub fn check_output(
    output: &Output<'_, impl Function, impl RegInfo>,
) -> Result<()>

Verifies the output of the register allocator.

If this fails then it indicates a bug in the register allocator, assuming the Function and RegInfo have passed validation.