regalloc3::entity::base
Trait EntityRef

pub trait EntityRef:
    Copy
    + Eq
    + ReservedValue {
    // Required methods
    fn new(index: usize) -> Self;
    fn index(self) -> usize;
}

A typed wrapper around an integer index.

Types implementing this trait can be used as keys for collections like PrimaryMap or SecondaryMap.

Required Methods

fn new(index: usize) -> Self

Creates a new entity reference from a raw index.

fn index(self) -> usize

Returns the index that was used to create this entity reference.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.