regalloc3
Module function

The Function trait which describes the input function.

This includes all details of the function including basic blocks, instructions and their operands.

Control flow graph and instructions

The allocator operates on an input program that is in a standard CFG representation: the function body is a sequence of basic blocks, and each block has a sequence of instructions and zero or more successors. The allocator also requires the client to provide predecessors for each block, and these must be consistent with the successors. The entry block may not have any predecessors.

The CFG must have no critical edges. A critical edge is an edge from block A to block B such that A has more than one successor and B has more than one predecessor.

Instructions are opaque to the allocator: their behavior is entirely described using a vector of Operands. Every block must end with a terminator instruction. The kind of terminator is determined by the number of successor blocks in a function:

A jump terminator has a single successor block and that successor block has more than one predecessor blocks.

A branch terminator has multiple successor blocks, each of which only has a single predecessor block.

A ret has no successor blocks and return from the function.

Both instructions and blocks are named by indices in contiguous index spaces. A block’s instructions must be a contiguous range of instruction indices, and block i’s first instruction must come immediately after block i-1’s last instruction.

Basic blocks in the CFG must be topologically sorted with regards to the dominator tree of the CFG. This means that if block A dominates block B then A must have a lower block index than B. Additionally, the heuristics used by the register allocator work best if loop bodies are continuous and properly nested. The best way to achieve this is to order block by a reverse-postorder (RPO) of the CFG, with loop back-edges prioritized to come earlier than other edges.

The first block is treated as the entry block of the CFG. It cannot have any predecessors and all other blocks must be reachable from it.

Operands and values

Every instruction operates on values by way of Operands. An operand consists of the 2 parts: the OperandKind and the OperandConstraint.

A Value represents an abstract value produced by an instruction and consumed by other instructions.

OperandKind describes what the operand is doing to a Value. This is usually one of two things: it defines (OperandKind::Def) a value as an output or it uses (OperandKind::Use) a value as an input.

OperandConstraint describes the constraints on which PhysReg can be assigned to an operand. This is usually OperandConstraint::Class which means that any register in a given RegClass can be selected. Some instructions (e.g. a function call) require values to be in a fixed register. For those, OperandConstraint::Fixed can be used.

SSA and block parameters

The function described by the Function trait must be in Static Single-Assignment (SSA) form. This means that each Value can only be defined once, by a single instruction. Additionally, because the CFG must be ordered by dominance, this means that the definition of a value must always appear before any uses of that value in the linear instruction order: the index of the defining Inst must be less than that of any instruction using the value.

At control-flow join points, a Value may need to represent the incoming value from multiple blocks, each of which has a separate definition. This is achieved using basic block parameters, which are equivalent to phi-nodes.

A basic block with multiple predecessor blocks may specify a set of block parameters with Function::block_params. This defines new Values at the start of the block. All predecessor blocks, which must end with a jump terminator, must provide a matching number of outgoing Values with Function::jump_blockparams. On entry to the block, the block parameter values will be initialized with the value from the block that control flow actually came from.

Reusing an input register for an output

Some instructions, particularly on ISAs like x86, only allow a single register to be specified as both an input and output, where the instruction clobber the original input value with the output value.

In the case of fixed registers, this can be represented by using the same fixed register for both an input and an output operand. However for class-constraint operands an explicit relationship needs to be created between the input and output operand to ensure they are assigned to the same register.

This is achieved by using OperandConstraint::Reuse for the output operand and specifying the index of the corresponding input operand, which must have OperandConstraint::Class. This will force the register allocator to assign to the output operand the same register as the designated input operand.

Non-allocatable registers

OperandKind::NonAllocatable can be used to specify a PhysReg that should be ignored by the register allocator and passed on to the output Allocation unmodified. This is intended to help with client that want to make use of special registers like the stack pointer or a hardware zero register without needing to change the number of operands in an instruction.

Any registers used with OperandKind::NonAllocatable must not be part of any register bank or register class.

Structs

Block

An opaque reference to a basic block in the input function.

Inst

An instruction index in the input function.

Operand

Information about an operand in an instruction.

Value

An opaque reference to an SSA value in the input function.

ValueGroup

A reference to a list of Values.

Enums

OperandConstraint

An OperandConstraint specifies which value should be assigned to this operand and any constraints on the location where this value should be placed.

OperandKind

The “kind” of an operand, which describes how an instruction makes use of an operand.

RematCost

Information about the cost of rematerializing a value into a register.

Constants

MAX_BLOCKS

Maximum number of basic blocks.

MAX_BLOCK_PARAMS

Maximum number of basic block parameters.

MAX_INSTS

Maximum number of instructions.

MAX_INST_OPERANDS

Maximum number of operands per instruction.

MAX_VALUES

Maximum number of SSA values.

Traits

Function

A trait defined by the register allocator client to provide access to its machine-instruction / CFG representation.

Type Aliases

InstRange

A range of instructions in the input function.