regalloc3::function
Trait Function

pub trait Function {

    // Required methods
    fn num_insts(&self) -> usize;
    fn num_blocks(&self) -> usize;
    fn block_insts(&self, block: Block) -> InstRange;
    fn inst_block(&self, inst: Inst) -> Block;
    fn block_succs(&self, block: Block) -> &[Block];
    fn block_preds(&self, block: Block) -> &[Block];
    fn block_immediate_dominator(&self, block: Block) -> Option<Block>;
    fn block_params(&self, block: Block) -> &[Value];
    fn inst_is_terminator(&self, inst: Inst) -> bool;
    fn jump_blockparams(&self, block: Block) -> &[Value];
    fn block_frequency(&self, block: Block) -> f32;
    fn inst_operands(&self, inst: Inst) -> &[Operand];
    fn inst_clobbers(&self, inst: Inst) -> &[RegUnit];
    fn num_values(&self) -> usize;
    fn value_bank(&self, value: Value) -> RegBank;
    fn num_value_groups(&self) -> usize;
    fn value_group_members(&self, group: ValueGroup) -> &[Value];
    fn can_rematerialize(&self, value: Value) -> Option<(RematCost, RegClass)>;
    fn can_eliminate_dead_inst(&self, inst: Inst) -> bool;

    // Provided methods
    fn insts(&self) -> Keys<Inst>  { ... }
    fn blocks(&self) -> Keys<Block>  { ... }
    fn block_dominates(&self, a: Block, b: Block) -> bool { ... }
    fn values(&self) -> Keys<Value>  { ... }
    fn value_groups(&self) -> Keys<ValueGroup>  { ... }

}

A trait defined by the register allocator client to provide access to its machine-instruction / CFG representation.

See the module-level documentation for more details.

Required Methods

fn num_insts(&self) -> usize

How many instructions are there?

fn num_blocks(&self) -> usize

How many blocks are there?

All blocks must be reachable from the entry block.

fn block_insts(&self, block: Block) -> InstRange

Provide the range of instruction indices contained in each block.

fn inst_block(&self, inst: Inst) -> Block

Returns the block containing the given instruction.

fn block_succs(&self, block: Block) -> &[Block]

Get CFG successors for a given block.

fn block_preds(&self, block: Block) -> &[Block]

Get the CFG predecessors for a given block.

fn block_immediate_dominator(&self, block: Block) -> Option<Block>

Returns the immediate dominator of the given block, or None if it is the entry block.

fn block_params(&self, block: Block) -> &[Value]

Get the block parameters for a given block.

Block parameters are only allowed on blocks with more than one predecessors.

fn inst_is_terminator(&self, inst: Inst) -> bool

Determine whether an instruction is an end-of-block branch or return.

All blocks must end with a terminator instruction, and terminator instructions are not allowed in the body of a block.

fn jump_blockparams(&self, block: Block) -> &[Value]

If block ends with a jump terminator, returns the outgoing block arguments.

Jump arguments are only allowed on blocks with a single successor.

The number of arguments must match the number incoming blockparams in the successor.

If a block has outgoing branch arguments then the terminator instruction cannot have any operands.

fn block_frequency(&self, block: Block) -> f32

Returns the estimated execution frequency of this block.

The allocator uses this to prefer placing moves in lower-frequency blocks and to prioritize registers for values that are used in higher- frequency blocks.

This number be be non-zero and positive. In general, good numbers to use are in the range of 10e-9 to 10e9 since this avoids issues floats overflowing into infinity and precision loss.

fn inst_operands(&self, inst: Inst) -> &[Operand]

Get the Operands for an instruction.

fn inst_clobbers(&self, inst: Inst) -> &[RegUnit]

Get the clobbers for an instruction; these are the registers units that, after the instruction has executed, hold values that are arbitrary, separately from the usual outputs to the instruction. It is invalid to read a register that has been clobbered; the register allocator is free to assume that clobbered registers are filled with garbage and available for reuse. It will avoid storing any value in a clobbered register that must be live across the instruction.

Another way of seeing this is that a clobber is equivalent to a Def of a fresh Value that is not used anywhere else in the program, with a fixed-register constraint that places it in a given PhysReg chosen by the client prior to regalloc.

Every register written by an instruction must either correspond to (be assigned to) an Operand of kind Def, or else must be a “clobber”.

This can be used to, for example, describe ABI-specified registers that are not preserved by a call instruction, or fixed physical registers written by an instruction but not used as an output, or fixed physical registers used as temps within an instruction out of necessity.

Duplicate clobbers are not allowed and clobbers may not overlap with any fixed-def operands on the same instruction.

fn num_values(&self) -> usize

Get the number of SSA values in use in this function.

fn value_bank(&self, value: Value) -> RegBank

The register bank associated with the given value.

A given value can only be associated with a single register bank, typically based on its type in the source IR. All operands using this value must have constraints from the same bank and it may only be linked to block parameters from the same bank.

fn num_value_groups(&self) -> usize

Get the number of value groups in use in this function.

fn value_group_members(&self, group: ValueGroup) -> &[Value]

Get the members of a value group.

fn can_rematerialize(&self, value: Value) -> Option<(RematCost, RegClass)>

Whether a Value can be re-materialized “cheaply” (specifically, cheaper than a stack spill and reload).

There are strict restrictions on rematerializable values: they cannot depend on any other allocatable register and they are only allowed to use a single destination register as scratch space.

Additionally, the register class for the rematerialization target must:

have an non-empty allocation order.

either allow spillsots or only have registers for which RegInfo::is_memory is false.

fn can_eliminate_dead_inst(&self, inst: Inst) -> bool

If all the outputs of an instruction are dead (never used), can the instruction be removed (i.e. it has no side effects apart from its outputs).

Instruction outputs can become dead due to rematerialization when all existing users are using the rematerialized value and the original instruction is no longer needed.

Provided Methods

fn insts(&self) -> Keys<Inst>

Iterator over all the Insts in this function.

fn blocks(&self) -> Keys<Block>

Iterator over all the Blocks in this function.

fn block_dominates(&self, a: Block, b: Block) -> bool

Returns whether block a dominates block b.

This should return true if a == b.

fn values(&self) -> Keys<Value>

Iterator over all the Values in this function.

fn value_groups(&self) -> Keys<ValueGroup>

Iterator over all the ValueGroups in this function.