regalloc3
Module output

Output of register allocation.

This is represented by the Output type and primarily consists of two parts: allocations and moves.

Allocations

For each instruction, the allocator will produce an array of Allocations, one per Operand. An allocation can represent either a SpillSlot or a PhysReg. Registers and spillslot are assigned according to the constraints on the operand, and will faithfully preserve the dataflow of the input program.

In the case of register group operands, only the first register of the group is returned in the allocation list.

Moves

In order to implement the necessary movement of data between allocations, the allocator needs to insert moves at various points between the original program instructions.

These moves may consist of reg-to-reg, reg-to-stack (spill) and stack-to-reg (fill) moves. Since most ISAs lack instructions for stack-to-stack moves, the register allocation will automatically split those using a scratch register in RegInfo::stack_to_stack_class.

Rematerialization

If a Value has been marked as being rematerializable then the allocator may insert an OutputInst::Rematerialize between two instructions which indicates that the given value should be rematerialized into the given Allocation. This is often cheaper than spilling to the stack, especially for constant values.

Structs

Allocation

An Allocation represents the end result of regalloc for an Operand.

Output

The output of the register allocator.

OutputIter

Iterator over the OutputInst of a block after register allocation.

SpillSlot

Space in the stackframe used by the allocator to temporarily store a value.

StackLayout

Positions of all the spill slots in the stack frame.

Enums

AllocationKind

Expanded form of Allocation as an enum.

OutputInst

Wrapper around either an original instruction or an inserted move.

Constants

MAX_SPILL_AREA_SIZE

Maximum size of the spill area.