regalloc3::output
Enum OutputInst

pub enum OutputInst<'a> {
    Inst {
        inst: Inst,
        operand_allocs: &'a [Allocation],
    },
    Rematerialize {
        value: Value,
        to: Allocation,
    },
    Move {
        from: Allocation,
        to: Allocation,
        value: Option<Value>,
    },
}

Wrapper around either an original instruction or an inserted move.

Variants

Inst

An original instruction, with its operands mapped to Allocations.

Note that due to rematerialization, not all instructions of the original function may be present. This can happen if the original instruction had no side effects and its outputs are not needed.

Fields

inst: Inst

The reference to the original instruction.

operand_allocs: &'a [Allocation]

The Allocation for each Operand of the instruction.

In the case of register group operands, only the allocation for the first group member is recorded here.

Rematerialize

A value which should be re-materialized into the given Allocation.

There are strict restrictions on rematerializable values: they cannot depend on any other allocatable register and they are only allowed to use a single destination register as scratch space.

Fields

value: Value

Value being rematerialized.

to: Allocation

Destination into which the rematerialized value should be wrriten.

Move

A move instruction inserted by the register allocator.

value indicates the original SSA value currently contained in the from allocation.

Fields

from: Allocation

Source location of the move.

to: Allocation

Destination location of the move.

value: Option<Value>

SSA value that is being copied. In the case of block parameter moves, this is the source value in the predecessor block.

This is None when a register is being temporarily evicted to a spill slot or restored from such a spill slot.