regalloc3
Module reginfo

The RegInfo trait which describes the registers and register classes in the target ISA.

The registers are described using a hierarchy of types:

A RegBank is a top-level grouping of registers, and contains one or more RegClass.

A RegClass represents a selection of PhysReg or RegGroup to allocate for a instruction operand.

A RegGroup represents a sequence of PhysReg which are allocated together for a single operand.

A PhysReg consists of one or more RegUnit which are the smallest unit of allocation.

Physical registers

A PhysReg is the unit of allocation that the register allocator works with. Fundamentally, the job of the register allocator is to assign a PhysReg (or a spill slot) to every use of a Value in a function.

Register banks

A RegBank represents a set of PhysReg between which values can be moved with move instructions. Register banks must not overlap: each PhysReg may only be part of a single register bank.

Each Value must be located in one register bank: all operand constraints related to this value must use registers and register classes from that register bank. Similarly, any pair of values linked together through tied operands or through block parameters must share the same register bank.

Register classes

A RegClass represents a subset of the registers of a RegBank which can be used by an instruction. Specifically, it describes which registers are allowed to satisfy a given constraint, and whether the constraint can be satisfied by stack spill slot (for instruction which can accept either a register or a memory operand).

Each register bank has an associated top-level register class. This is a register class which covers the entire set of registers in the bank. Top-level register classes are required to allow moves to/from spillslots.

Each register bank also has an associated stack-to-stack register class. This is used when a scratch register is required to resolve a move between 2 memory locations.

Register classes may overlap either completely or partially with other register classes in the same bank.

Allocation order

Each register class has an allocation order which is the order in which the register allocator will attempt to select a suitable register for a constraint of this class.

The RegInfo trait specifies the allocation order as 4 sets of registers:

AllocationOrderSet::Preferred

AllocationOrderSet::NonPreferred

AllocationOrderSet::CalleeSavedPreferred

AllocationOrderSet::CalleeSavedNonPreferred

Preferred registers generally have a more compact or efficient encoding than non-preferred registers. Callee-saved registers are treated as normal registers if they have already been used at least once in the function, and are otherwise given the lowest priority. This minimizes the need for saving and restoring these registers on function entry and exit.

For a minimal implementation, it is possible to only specify registers as AllocationOrderSet::Preferred and leave the other sets empty.

Any registers that are members of a class but not in the allocation order are only selected by the register allocator if it helps to satisfy a Fixed register constraint and are otherwise never used.

Fixed stack slots

A PhysReg can also be used to represent a memory location instead of a register. This is useful to model stack slots for function arguments and return values which

Since it is undesirable to use these stack slots unless specifically needed, these can be made part of a register class without being part of its allocation order. This means that these PhysReg will only be selected if necessary to satisfy a Fixed constraint, but not a Class constraint.

Compound registers

Some architectures may have registers that overlap: for example on AArch32 the D0 64-bit register overlaps with the S0 and S1 32-bit registers. This is modeled in the register allocator by using “register units”: S0 and S1 can be defined as covering single register units unit0 and unit1, while D0 can be defined as a compound register covering both unit0 and unit1.

Liveness is only tracked at the register unit level, so the register units must completely cover all the bits of the compound register.

Overlaps are only allowed with registers in other banks (i.e. D0 and S0 must be in different register banks). Within a bank no registers may share a unit.

Compound registers are still limited to holding a single Value. They only interact with register units to avoid conflicts with values held in other registers than they share a register unit with.

Register groups

Some instructions take a sequence of registers as operands, where only the first register is encoded in the instruction (e.g. LD4 on AArch64). This cannot be modeled using compound registers because each register needs to be assigned a different Value.

The register allocator can model this using register groups, where a RegGroup represents a sequence of PhysReg. Register classes are generalized to contain either a set of PhysReg or a set of RegGroup, with the restriction that all register groups in a class must have the same number of registers.

Instruction operands can then be specified to require a group of values using OperandKind::UseGroup, OperandKind::DefGroup or OperandKind::EarlyDefGroup. These operands take a ValueGroup which represents a list of Values.

Group operands can only be used with OperandConstraint::Class and the class must be a class of register groups where groups have the same number of registers as the number of Values in the ValueGroup.

The Allocation in OutputInst::Inst::operand_allocs will only contain a single entry for the operand which holds the first register of the register group that was allocated for the operand.

Structs

PhysReg

A location to which a value can be mapped to.

RegBank

A set of registers between which values can be copied with move instructions.

RegClass

A set of PhysReg or RegGroup which can be allocated for an operand.

RegGroup

A sequence of multiple registers which are assigned together for a single operand.

RegOrRegGroup

A sequence of registers which are assigned together for a single operand.

RegUnit

A subset of a register which cannot be further split.

SpillSlotSize

The size of a spillslot, which must be a power of two.

Enums

AllocationOrderSet

Sets of registers in a RegClass grouped by their priority for register allocation.

Constants

MAX_GROUP_SIZE

Maximum size of a register group.

MAX_PHYSREGS

Maximum number of registers.

MAX_REG_BANKS

Maximum number of register banks.

MAX_REG_CLASSES

Maximum number of register classes.

MAX_REG_GROUPS

Maximum number of register groups.

MAX_REG_UNITS

Maximum number of register units.

MAX_UNITS_PER_REG

Maximum number of register units in a compound register.

Traits

RegInfo

Trait which describes the physical registers and register classes that may be allocated from.

Type Aliases

PhysRegSet

A set of PhysReg encoded as a fixed-size bit set.

RegClassSet

A set of RegClass encoded as a fixed-size bit set.

RegOrRegGroupSet

A set of RegOrRegGroup encoded as a fixed-size bit set.

RegUnitSet

A set of RegUnit encoded as a fixed-size bit set.