regalloc3::reginfo
Trait RegInfo

pub trait RegInfo {

    // Required methods
    fn num_banks(&self) -> usize;
    fn top_level_class(&self, bank: RegBank) -> RegClass;
    fn stack_to_stack_class(&self, bank: RegBank) -> RegClass;
    fn bank_for_class(&self, class: RegClass) -> RegBank;
    fn bank_for_reg(&self, reg: PhysReg) -> Option<RegBank>;
    fn spillslot_size(&self, bank: RegBank) -> SpillSlotSize;
    fn num_classes(&self) -> usize;
    fn class_members(&self, class: RegClass) -> RegOrRegGroupSet;
    fn class_includes_spillslots(&self, class: RegClass) -> bool;
    fn class_spill_cost(&self, class: RegClass) -> f32;
    fn allocation_order(
        &self,
        class: RegClass,
        set: AllocationOrderSet,
    ) -> &[RegOrRegGroup];
    fn sub_classes(&self, class: RegClass) -> RegClassSet;
    fn class_group_size(&self, class: RegClass) -> usize;
    fn num_regs(&self) -> usize;
    fn reg_units(&self, reg: PhysReg) -> &[RegUnit];
    fn is_memory(&self, reg: PhysReg) -> bool;
    fn num_reg_groups(&self) -> usize;
    fn reg_group_members(&self, group: RegGroup) -> &[PhysReg];
    fn group_for_reg(
        &self,
        reg: PhysReg,
        group_index: usize,
        class: RegClass,
    ) -> Option<RegGroup>;

    // Provided methods
    fn banks(&self) -> Keys<RegBank>  { ... }
    fn classes(&self) -> Keys<RegClass>  { ... }
    fn common_subclass(&self, a: RegClass, b: RegClass) -> Option<RegClass> { ... }
    fn regs(&self) -> Keys<PhysReg>  { ... }
    fn reg_groups(&self) -> Keys<RegGroup>  { ... }

}

Trait which describes the physical registers and register classes that may be allocated from.

This is separate from the Function trait since this data is usually independent of the function currently being compiled.

See the module-level documentation for more details.

Required Methods

fn num_banks(&self) -> usize

Get the number of RegBank in use.

fn top_level_class(&self, bank: RegBank) -> RegClass

Returns the top-level register class of the given register bank.

This class encompasses all of the register in the bank, and is required to accept spillslot allocations. It is a superclass of all other classes in this bank.

fn stack_to_stack_class(&self, bank: RegBank) -> RegClass

Returns the class from which to select scratch registers for splitting stack-to-stack moves.

Move instructions on most ISAs cannot perform a memory-to-memory copy, so such moves must be split into 2 halves using an intermediate scratch register from this class.

The returned class must not be a group class and cannot contain any register for which RegInfo::is_memory is true.

fn bank_for_class(&self, class: RegClass) -> RegBank

Returns the register bank that the given RegClass is a member of.

fn bank_for_reg(&self, reg: PhysReg) -> Option<RegBank>

Returns the register bank that the given PhysReg is a member of.

Returns None for non-allocatable registers.

fn spillslot_size(&self, bank: RegBank) -> SpillSlotSize

Spillslot size needed for a value in this register bank.

The spillslot is guaranteed to be aligned to this size.

fn num_classes(&self) -> usize

Get the number of RegClass in use.

fn class_members(&self, class: RegClass) -> RegOrRegGroupSet

Returns the set of PhysReg or RegGroup contained within a register class.

This indicates that reg may be assigned to an operand constrainted to class. This is the case even if the register is not a part of the allocation order.

fn class_includes_spillslots(&self, class: RegClass) -> bool

Whether an operand constrainted to this register class can be allocated to a SpillSlot instead of a PhysReg.

This must return true for top-level register classes, and false for register group classes.

fn class_spill_cost(&self, class: RegClass) -> f32

The spill cost of a class is defined as the cost that needs to be paid if a value constrained to this class is spilled to the stack instead of allocated to a register.

For most register classes, a default value of 1.0 can be used which indicates a cost of a single spill/reload instruction. In fact, there is no reason to use any other value if the class does not allow direct use of spillslots.

For register classes that can be allocated to spillslots, the value should represent the cost of accessing a stack slot compared to a real register. A good value is aroud 0.5, which indicates a cost that is not as big as a full spill/reload, but still

A spill cost of 0.0 indicates that there is no cost to spilling the value to the stack. This is useful for stack constraints (e.g. used for GC roots) or for values that are only read by trap handlers.

fn allocation_order( &self, class: RegClass, set: AllocationOrderSet, ) -> &[RegOrRegGroup]

Returns a set of PhysReg or RegGroup to try allocating for an operand constrained to the given register class.

Several sets of registers can be provided with different priority levels as determined by the AllocationOrderSet enum. The register allocator will attempt to select an available registers by probing the sets in this order:

Preferred registers.

Preferred callee-saved registers that have already been used in the function.

Non-preferred registers.

Non-preferred callee-saved registers that have already been used in the function.

Preferred callee-saved registers that have not yet been used in the function.

Non-preferred callee-saved registers that have not yet been used in the function.

This arrangement prioritizes registers with smaller encodings and penalizes registers that need to be saved on function entry and restored on function exit (but only if that cost hasn’t already been paid).

All registers (or register groups) returned by this function must be a member of the register class. However not all members need to be in the allocation order. Registers outside the allocation order will only be selected to satisfy a fixed-register operand constraint. This is useful for “fake” registers such as fixed stack slots which are slower to access than a register.

fn sub_classes(&self, class: RegClass) -> RegClassSet

Returns the set of sub-classes of class, including itself.

A sub-class must be from the same register bank as its superclass, and must contain a strict subset of the registers in the superclass.

A class of register groups can only have other group classes of the same group size as sub-classes. However it may have a non-group class as a superclass as long as

Classes must be topologically ordered: sub-classes must have a higher index than their superclass. Additionally, all register classes in a register bank must be a subclass of the top-level class of that bank.

fn class_group_size(&self, class: RegClass) -> usize

The group size of registers in this class.

If this returns a value greater than 1 then all PhysReg in this register class must be register groups with the given number of members, with each member coming from the register bank of this class.

fn num_regs(&self) -> usize

Get the number of PhysReg in use.

fn reg_units(&self, reg: PhysReg) -> &[RegUnit]

Returns the set of register units covered by a register.

See the module-level documentation for more details.

This is ignored for non-allocatable registers used with OperandKind::NonAllocatable. It can be an empty list in such cases.

fn is_memory(&self, reg: PhysReg) -> bool

Returns whether this PhysReg represents a memory location (e.g. a fixed stack slot) instead of a register.

Allocation::is_memory will forward to this function when the Allocation represents a PhysReg.

This has 2 effects on register allocation:

Memory to memory moves are resolved by using a temporary scratch register since such more are not supported by hardware.

Rematerialization can be more aggressive when it can avoid a load from memory, depending on RematCost.

fn num_reg_groups(&self) -> usize

Get the number of RegGroup in use.

fn reg_group_members(&self, group: RegGroup) -> &[PhysReg]

Returns the sequence of registers for a RegGroup.

fn group_for_reg( &self, reg: PhysReg, group_index: usize, class: RegClass, ) -> Option<RegGroup>

Searches members of class for a RegGroup which contains reg at position group_index within the group.

This can only be called for register classes with a group size greater than 1.

Provided Methods

fn banks(&self) -> Keys<RegBank>

Iterator over all the RegBanks.

fn classes(&self) -> Keys<RegClass>

Iterator over all RegClasses.

fn common_subclass(&self, a: RegClass, b: RegClass) -> Option<RegClass>

Computes the largest common sub-class of the two register classes, or None if no common sub-class exists.

fn regs(&self) -> Keys<PhysReg>

Iterator over all PhysRegs.

fn reg_groups(&self) -> Keys<RegGroup>

Iterator over all RegGroups.